requirements.txt

---

### **`requirements.txt`**
```txt
# Core Libraries
torch>=2.0.0
torchvision>=0.15.0
streamlit>=1.25.0
Pillow>=9.0.0
numpy>=1.24.0